netrecon/requirements.txt
flask==3.0.3
flask-dance==7.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-whois==0.9.4
dnspython==2.6.1
Werkzeug==3.0.3
oauthlib==3.2.2
requests-oauthlib==2.0.0
SQLAlchemy==2.0.31
